added check for riskAnalysis
